wip: add transaction and other fixing money account req res

+++ front-end/Takana/app/src/main/java/com/example/takana/presentation/money_account/MoneyAccountAddEditActivity.kt
@@ -84,7 +84,7 @@ class MoneyAccountAddEditActivity : AppCompatActivity() {
                         etAccountAmount.text.toString().toLong(),
                         accountTypeId,
                         user?.userId!!,
-                        etBankAccountNumber.text.toString().toInt()
+                        etBankAccountNumber.text.toString()
                     )
                     viewModelDeleteDataMoneyAccount()
                 }
@@ -96,7 +96,7 @@ class MoneyAccountAddEditActivity : AppCompatActivity() {
                         etAccountAmount.text.toString().toLong(),
                         accountTypeId,
                         user?.userId!!,
-                        etBankAccountNumber.text.toString().toInt()
+                        etBankAccountNumber.text.toString()
                     )
                     viewModelUpdateDataMoneyAccount()
                 }
